// SDK/projects_api/client/models/LeagueData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { League } from './League';
import type { NewEntries } from './NewEntries';
import type { Standings } from './Standings';
export type LeagueData = {
    new_entries: NewEntries;
    last_updated_data: string;
    league: League;
    standings: Standings;
};