// SDK/projects_api/client/models/LeagueCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LeaguePrizeCreate } from './LeaguePrizeCreate';
export type LeagueCreate = {
    name: string;
    description?: (string | null);
    external_id_fpl: number;
    prizes?: (Array<LeaguePrizeCreate> | null);
};